Updates
Had to change the time zone back to eastern since all the games start
times are in eastern.

Reformatted message output

Changed the way it checks for in progress games.

Adjusted games table schema for games that are not started yet.

Changed created and modified dates to be date  times.

Removed unnecessary league table code